1118：run_ft_enhanced_culturemoe_gen_fixed.sh,run_eval_enhanced_culturemoe_from_components.sh，run_eval_lora_only_from_components.sh; 增加moe；8

+++ src/llamafactory/model/enhanced_culturemoe_fixed.py
@@ -138,8 +138,8 @@ class EnhancedCultureMoEFixed(LlamaSharedRouterExpertsModel):
         self.cultural_gate = self.cultural_gate.to(dtype=base_dtype)
 
         # 确保参数也使用正确的数据类型
-        self.culture_loss_alpha_enhanced = self.culture_loss_alpha_enhanced.to(dtype=base_dtype)
-        self.culture_loss_beta_enhanced = self.culture_loss_beta_enhanced.to(dtype=base_dtype)
+        self.culture_loss_alpha_enhanced.data = self.culture_loss_alpha_enhanced.data.to(dtype=base_dtype)
+        self.culture_loss_beta_enhanced.data = self.culture_loss_beta_enhanced.data.to(dtype=base_dtype)
 
         logging.info(f"All components converted to dtype: {base_dtype}")
